requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptg==0.4.0
cryptography==42.0.3
databases==0.8.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.40.1
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
mautrix==0.20.0
multidict==6.0.4
numpy==1.24.4
orjson==3.9.13
passlib==1.7.4
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
Telethon==1.28.5
TgCrypto==1.2.5
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
